1.19 update.

+++ src/MiNET/MiNET/Particles/DestroyBlockParticle.cs
@@ -33,7 +33,7 @@ namespace MiNET.Particles
 	{
 		public DestroyBlockParticle(Level level, Block block) : base(0, level)
 		{
-			Data = (int) block.GetRuntimeId();
+			Data =  block.GetRuntimeId();
 			Position = block.Coordinates;
 		}
